Move memorystore to its own file
