+++ store/memorystore_test.go
